Move store to separate module
